[trading] Make position size integer and avoid duplicate buys
[Description]
- Fixed quota debug message formatting in market/dataprovider.py
- Updated Trading class to skip buy evaluation for held positions
- Changed position size fields to integer type across trading modules:
 - Transaction table size column (REAL → INTEGER)
 - Trade and Recorder classes (size → int)
 - Trade record printing headers (size → Size, commission → Commission)
- Added position checks in Evaluate class and test scenarios
- Properly type-cast database size values to integers

[Motivation]
- Ensure position sizes are integers to avoid fractional shares
- Prevent duplicate buys for already-held positions
- Fix test scenario implementation after logic changes
- Resolve potential data integrity issues with float sizes
- Improve trade table readability with consistent formatting

[Self-Test]
- Trading unit tests passed (testutility/trading.py)
- DB schema migration validation for size column change

+++ trading/tradedata.py
@@ -31,7 +31,7 @@ class Database(uDatabase):
                         trade_id TEXT NOT NULL,
                         action TEXT NOT NULL,
                         price REAL NOT NULL,
-                        size REAL NOT NULL,
+                        size INTEGER NOT NULL,
                         timestamp INTEGER NOT NULL,
                         commission REAL,
                         FOREIGN KEY(trade_id) REFERENCES Trades(trade_id)
